move pi and anthorpic into dirs under harness
