+++ src/lib/agent/runner/harness/pi/__tests__/security.test.ts
@@ -3,7 +3,7 @@ import {
   createSecurityExtension,
   MAX_TOOL_CALLS,
   type PiExtensionApiLike,
-} from '../pi-security';
+} from '../security';
 
 const block = (toolName: string, input: Record<string, unknown>) =>
   evaluateToolCall(toolName, input).block;
