+++ src/lib/agent/runner/harness/pi/mcp.ts
